Crud pessoas

+++ app/database/models/pessoas.py
@@ -1,5 +1,5 @@
 from app.database.db import Base
-from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy import Column, Integer, String, DateTime, BigInteger
 from datetime import datetime
 from datetime import datetime, timezone
 
@@ -9,6 +9,6 @@ class Pessoa(Base):
     id = Column(Integer, primary_key=True, index=True)
     nome = Column(String)
     email = Column(String, unique=True)
-    telefone = Column(Integer, unique=True, )
+    telefone = Column(BigInteger, unique=True, )
     createdAt = Column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc))
     updatedAt = Column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc), onupdate=lambda: datetime.now(timezone.utc))
